+++ index.html
@@ -102,8 +102,9 @@
 				</div>
 			</div>
 		</div>
-		<script type="module" src="scripts/models.js"></script>
+		<script type="module" src="scripts/api.js"></script>
 		<script type="module" src="scripts/views.js"></script>
+		<script type="module" src="scripts/controllers.js"></script>
 		<script type="module" src="scripts/script.js"></script>
 	</body>
 </html>
