+++ scripts/api.js
@@ -1,28 +1,29 @@
 const baseURL = "http://localhost:8000/api/v1/titles/?format=json&sort_by=-imdb_score&genre="
 
+async function buildMovieList(json, movies, quantity){
+    for (let result of await json.results){
+        if (movies.length < quantity){
+            movies.push(result);
+        }
+    }
+}
+
 /**
- * 
- * @param {string} genre 
- * @param {number} quantity 
+ * Makes a list with objects fetched from API server.
+ * Depends of genre and number of films required.
+ * @param {string} genre: If genre is an empty string, API will return best films in all categories.
+ * @param {number} quantity: quantity of films desired in the final list.
  * @returns a list of movies
  */
 export async function fetchMovies(genre, quantity){
     let movies = []
     let response = await fetch(baseURL + genre);
     let json = await response.json();
-    for (let result of await json.results){
-        if (movies.length < quantity){
-            movies.push(result);
-        }
-    }
+    buildMovieList(json, movies, quantity);
     while (movies.length < quantity){
         response = await fetch(await json.next);
         json = await response.json();
-        for (let result of await json.results){
-            if (movies.length < quantity){
-                movies.push(result);
-            }
-        }
+        buildMovieList(json, movies, quantity);
     }
     return movies;
 }
